// src/types/form.ts
export interface FormData {
  productDescription: string;
  name: string;
  qualifications: string;
  avatarName: string;
  pronoun: string;
  personalPronoun: string;
  possessivePronoun: string;
  currentIdentity: string;
  currentIdentityPlural: string;
  idealIdentity: string;
  focus1: string;
  focus2: string;
  desire1: string;
  desire2: string;
  immediateResult1: string;
  immediateResult2: string;
  immediateResult3: string;
  enemies: string;
  problem1: string;
  problem2: string;
  problem3: string;
  pain1: string;
  pain2: string;
  pain3: string;
  question1: string;
  question2: string;
  question3: string;
  roadblock1: string;
  roadblock2: string;
  roadblock3: string;
  bottomLine1: string;
  bottomLine2: string;
  bottomLine3: string;
  productTitle: string;
  productType: string;
  benefit1: string;
  benefit2: string;
  benefit3: string;
  result1: string;
  painToAvoid: string;
}